[feature/detail] 디테일뷰모델 : 쉐어드레포지토리로 변경

+++ shared/src/commonMain/kotlin/com/kmm/moonflower/feature/garden/data/local/GardenPlantingLocalDataSource.kt
@@ -11,7 +11,7 @@ interface GardenPlantingLocalDataSource {
 
     suspend fun getPlantedGardens(): List<PlantAndGardenPlantings>
 
-    suspend fun insertGardenPlanting(plantId: String, plantDate: Int, lastWateringDate: Int)
+    suspend fun insertGardenPlanting(plantId: String, plantDate: Long, lastWateringDate: Long)
 
     suspend fun deleteGardenPlanting(gardenId : Int)
 }
